% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MTTKFit.R
\name{fitInfo}
\alias{fitInfo}
\alias{fitInfo,MTTKFit-method}
\title{Fit Metadata}
\usage{
fitInfo(x)

\S4method{fitInfo}{MTTKFit}(x)
}
\arguments{
\item{x}{An \code{MTTKFit}.}
}
\value{
A named list.
}
\description{
\code{fitInfo()} returns the metadata list stored alongside an \code{MTTKFit}.
}
