% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MTTKExperiment-accessors.R
\name{genomeData}
\alias{genomeData}
\alias{genomeData<-}
\alias{genomeData,MTTKExperiment-method}
\alias{genomeData<-,MTTKExperiment-method}
\title{Genome-Level Metadata}
\usage{
genomeData(x)

genomeData(x) <- value

\S4method{genomeData}{MTTKExperiment}(x)

\S4method{genomeData}{MTTKExperiment}(x) <- value
}
\arguments{
\item{x}{An \code{MTTKExperiment}.}

\item{value}{For replacement methods, an \code{S4Vectors::DataFrame},
\code{data.frame}, or \code{NULL}.}
}
\value{
\code{genomeData()} returns an \code{S4Vectors::DataFrame}. The replacement
method returns an updated \code{MTTKExperiment}.
}
\description{
\code{genomeData()} returns the genome-level metadata stored in
\code{rowData(genomeExperiment(x))}.
}
