% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MTTKExperiment-accessors.R
\name{genomeTree}
\alias{genomeTree}
\alias{genomeTree<-}
\alias{genomeTree,MTTKExperiment-method}
\alias{genomeTree<-,MTTKExperiment-method}
\title{Genome-Level Phylogeny}
\usage{
genomeTree(x)

genomeTree(x) <- value

\S4method{genomeTree}{MTTKExperiment}(x)

\S4method{genomeTree}{MTTKExperiment}(x) <- value
}
\arguments{
\item{x}{An \code{MTTKExperiment}.}

\item{value}{For replacement methods, an \code{ape::phylo} object or \code{NULL}.}
}
\value{
\code{genomeTree()} returns an \code{ape::phylo} object or \code{NULL}. The
replacement method returns an updated \code{MTTKExperiment}.
}
\description{
\code{genomeTree()} returns the genome phylogeny stored alongside the
genome-level layer of an \code{MTTKExperiment}.
}
\details{
The returned tree should have one tip per genome in \code{genomeExperiment(x)},
with tip labels that match \code{rownames(genomeExperiment(x))}. Rooted trees are
recommended in general and required for clade-scanning workflows.
}
