% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MTTKFit.R
\name{modelObjects}
\alias{modelObjects}
\alias{modelObjects,MTTKFit-method}
\title{Stored Model Objects}
\usage{
modelObjects(x)

\S4method{modelObjects}{MTTKFit}(x)
}
\arguments{
\item{x}{An \code{MTTKFit}.}
}
\value{
A named list.
}
\description{
\code{modelObjects()} returns the optional backend model objects stored in an
\code{MTTKFit}.
}
