% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTTKExample.R
\docType{data}
\name{MTTKExample}
\alias{MTTKExample}
\title{Example \code{MTTKExperiment}}
\format{
An \code{MTTKExperiment} object with 6 rows and 4 columns.
}
\source{
Generated from \code{data-raw/make-example-data.R}.
}
\usage{
MTTKExample
}
\description{
\code{MTTKExample} is a packaged \code{MTTKExperiment} object for examples,
documentation, and quick interactive use.
}
\details{
The object contains two assays (\code{rna_counts} and \code{dna_counts}), six genes,
four samples, three genomes, and a small set of biological and functional
link tables.
}
\examples{
data("MTTKExample")
MTTKExample
genomeData(MTTKExample)

}
\keyword{datasets}
