% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTTKExample.R
\docType{data}
\name{MTTKShowcase}
\alias{MTTKShowcase}
\title{Showcase \code{MTTKExperiment}}
\format{
An \code{MTTKExperiment} object with 16 rows and 6 columns.
}
\source{
Generated from \code{data-raw/make-example-data.R}.
}
\usage{
MTTKShowcase
}
\description{
\code{MTTKShowcase} is a packaged \code{MTTKExperiment} intended for vignette-style
walkthroughs of the main mttk workflows.
}
\details{
The object contains one gene-level assay (\code{rna_gene_counts}), two
genome-level assays (\code{rna_genome_counts} and \code{dna_genome_counts}), a rooted
genome phylogeny, sixteen genes, six samples, four genomes, and a small set
of overlapping functional mappings chosen to support KO-, module-,
pathway-, genome-, and group-comparison examples.
}
\examples{
data("MTTKShowcase")
MTTKShowcase
genomeData(MTTKShowcase)
genomeTree(MTTKShowcase)

}
\keyword{datasets}
